==> 01. Creating unique identifiers.ipynb <==
import numpy as np
import pandas as pd
import xlsxwriter

SU = pd.read_csv("../tables/Survey Units.txt", header=None)
exSU = pd.read_excel('../DataFiles/Survey Units.xlsx')

SU.columns = exSU.columns

su = SU[['SR', 'Begin ARM', 'End ARM']]
su.columns = ['SR', 'BA', 'EA']

su.head()

sr = np.unique(np.array(su.SR))

list(su[su.SR == 2].min())[1]

df = pd.DataFrame(sr)

df.columns = ['SR']

def addMin(row):
    sr = row['SR']
    return list(su[su.SR == sr].min())[1]
def addMax(row):
    sr = row['SR']
    return list(su[su.SR == sr].max())[2]

df.loc[:, 'Begins'] = df.apply(addMin, axis = 1)
df.loc[:, 'Ends'] = df.apply(addMax, axis = 1)

df.head()

d1 = pd.DataFrame(columns=('SR', 'bs', 'es'))

for i in range(5):
    d1.loc[i] = [df.loc[i, 'SR'], list(df[df.SR ==(i+2)].Begins)[0], df.loc[i, 'Begins']+0.1]

df[df.SR ==3].Begins

d1

db = pd.DataFrame(columns=('SR', 'bs', 'es'))
for road in np.array(df.SR):
    d1 =  pd.DataFrame(columns=('SR', 'bs', 'es'))
    ns = int(df[df.SR == road].Ends/0.1) + 1
    for i in range(ns):
        if i == 0:
            d1.loc[i] = [road, list(df[df.SR == road].Begins)[0]*10, list(df[df.SR == road].Begins)[0]*10 + 1]
        else:
            d1.loc[i] = [road, d1.loc[(i-1), 'es'] , d1.loc[(i-1), 'es'] + 1]
    db = pd.concat([db,d1])

print np.sum(np.array(df[['Ends']])*10)
print len(db)

db.loc[:, 'UID'] = np.arange(len(db))+1

db.head()

db.to_csv('../FinalDatasets/sid.csv')

def getUID(row):
    rd = row['SR']
    bg = row['bs']
    end = row['es']
    res = db[((db.SR==rd) & ((db.bs == bg) & (db.es == end)))].UID
    if list(res) == []:
        return None
    else:
        return list(res)[0]

su.head()

def roundme(n):
    return int(n*10)

su.loc[:, 'bs'] = su.loc[:, 'BA'].apply(roundme)
su.loc[:, 'es'] = su.loc[:, 'EA'].apply(roundme)

su.head(10)

su.loc[:, 'UID'] = su.apply(getUID, axis = 1)

print len(su[np.isnan(su.UID)])
print len(su)

su_wid = su[~np.isnan(su.UID)]
su2 = su[np.isnan(su.UID)].sort(columns = ['SR', 'BA', 'EA'])

print len(su_wid)
print len(su2)

# su2.head()

su_ids = np.array(su_wid.UID)

# (su2.loc[4064, 'EA'] - su2.loc[4064, 'BA'] > 0.05)  & (su2.loc[4064, 'EA'] - su2.loc[4064, 'BA'] < 0.11)

def fillUID(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    begin = row['bs']
    if (end - bg > 0.05) & (end - bg < 0.15):
        res = db[((db.SR==rd) & (db.bs == begin))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in su_ids: #  change su_ids to a list and just append x to it if not found in ids.
                return x
            return None

su2.loc[:, 'UID'] = su2.apply(fillUID, axis = 1)

len(su2[np.isnan(su2.UID)])

su_w2 = su2[~np.isnan(su2.UID)]
su3 = su2[np.isnan(su2.UID)]

su_ids = np.concatenate((su_ids,np.array(su_w2.UID)),axis=0)

def fillUID2(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    begin = row['bs']
    if (end - bg > 0.01):
        res = db[((db.SR==rd) & (db.bs == int((bg*10 + end*10)/2)))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in su_ids:
                return x
            return None

su3.loc[:, 'UID'] = su3.apply(fillUID2, axis = 1)

temp = su3.dropna(subset = ['UID'])
print len(temp)

su_w3 = su3[~np.isnan(su3.UID)]
su4 = su3[np.isnan(su3.UID)]

su_ids = np.concatenate((su_ids,np.array(su_w3.UID)),axis=0)

print len(np.unique(su_ids))
print len(db)
print len(su)

su_final = pd.concat([su_wid, su_w2, su_w3])

su_ds = su_final.drop_duplicates(subset = ['UID'])

print len(su_ds)

SU.head()

SU_FINAL = pd.merge(su_ds, SU, how = 'left', left_on = ['SR', 'BA', 'EA'], right_on = ['SR', 'Begin ARM', 'End ARM'])

SU2 = SU_FINAL.dropna(subset = ['Lane Miles'])

SU_FINAL.to_csv('../FinalDatasets/SurveyUnits.csv')

print len(SU2), len(SU_FINAL)

# Contracts dataset

excd= pd.read_excel("../DataFiles/Contracts.xlsx")

cd= pd.read_csv("../tables/Contracts.txt", header = None)
cd.columns = excd.columns

cds = cd[['SR', 'Begin ARM', 'End ARM']]
cds.columns = ['SR', 'BA', 'EA']
cds.loc[:, 'bs'] = cds.loc[:, 'BA'].apply(roundme)
cds.loc[:, 'es'] = cds.loc[:, 'EA'].apply(roundme)
# cdP = cds[~np.isnan(cds.UID)]

cds.loc[:, 'UID'] = cds.apply(getUID, axis = 1)

cds_wid = cds[~np.isnan(cds.UID)]
print len(cds_wid)
cds2 = cds[np.isnan(cds.UID)].sort(columns = ['SR', 'BA', 'EA'])
print len(cds2)

cds_ids = list(np.array(cds_wid.UID))

def fillUID(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    begin = row['bs']
    if (end - bg < 0.15):
        res = db[((db.SR==rd) & (db.bs == begin))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in cds_ids: #  change su_ids to a list and just append x to it if not found in ids.
                cds_ids.append(x)
                return x
            return None

cds2.loc[:, 'UID'] = cds2.apply(fillUID, axis = 1)

cds_w2 = cds2[~np.isnan(cds2.UID)]
print len(cds_w2)
cds3 = cds2[np.isnan(cds2.UID)]
print len(cds3)

print len(cds_ids) - len(cds_wid)

def fillUID2(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    begin = row['bs']
    if (end - bg > 0.15):
        res = db[((db.SR==rd) & (db.bs == int((bg*10 + end*10)/2)))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in cds_ids:
                cds_ids.append(x)
                return x
            return None

cds3.loc[:, 'UID'] = cds3.apply(fillUID2, axis = 1)

cds_w3 = cds3[~np.isnan(cds3.UID)]
print len(cds_w3)
cds4 = cds3[np.isnan(cds3.UID)]
print len(cds4)

print len(np.unique(cds_ids))
print len(db)
print len(cds)

def fillUID3(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    endARM = row['es']
    if (end - bg > 0.05) & (end - bg < 0.5):
        res = db[((db.SR==rd) & (db.es == endARM))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in cds_ids:
                cds_ids.append(x)
                return x
            return None

cds4.loc[:, 'UID'] = cds4.apply(fillUID3, axis = 1)

cds_w4 = cds4[~np.isnan(cds4.UID)]
print len(cds_w4)
cds5 = cds4[np.isnan(cds4.UID)]
print len(cds5)

def fillUID4(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
#     endARM = row['es']
    if (end - bg > 0.25):
        res = db[((db.SR==rd) & (db.bs == int((bg*10 + end*10)/3)))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in cds_ids:
                cds_ids.append(x)
                return x
            return None
def fillUID5(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
#     endARM = row['es']
    if (end - bg > 0.25):
        res = db[((db.SR==rd) & (db.es == int((bg*10 + end*10)/3)))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in cds_ids:
                cds_ids.append(x)
                return x
            return None

cds5.loc[:, 'UID'] = cds5.apply(fillUID4, axis = 1)

cds_w5 = cds5[~np.isnan(cds5.UID)]
print len(cds_w5)
cds6 = cds5[np.isnan(cds5.UID)]
print len(cds6)

cds6.loc[:, 'UID'] = cds6.apply(fillUID5, axis = 1)

cds_w6 = cds6[~np.isnan(cds6.UID)]
print len(cds_w6)
cds7 = cds6[np.isnan(cds6.UID)]
print len(cds7)

cds_all = pd.concat([cds_wid, cds_w2, cds_w3, cds_w4, cds_w5, cds_w6])

cds_all = cds_all.drop_duplicates(subset = ['UID'])

cds_final = pd.merge(cds_all, cd, how = 'left', left_on = ['SR', 'BA', 'EA'], right_on = ['SR', 'Begin ARM', 'End ARM'])
cds_FINAL = cds_final.dropna(subset = ['Surface Thickness'])
CDS = cds_FINAL.drop_duplicates(subset = ['SR', 'Begin ARM', 'End ARM', 'Construction End Date'])

print len(np.unique(cds_ids)), len(CDS), len(cd)

CDS[['SR', 'Begin ARM', 'End ARM', 'UID', 'Construction End Date', 'ID']].head()

CDS.to_csv('../FInalDatasets/Contracts.csv')

# Contracts History

exCH= pd.read_excel("../DataFiles/Contract History.xlsx")
CH = pd.read_csv('../tables/Contract History.txt', header = None)

CH.columns = exCH.columns
ch = CH[['SR', 'Begin ARM', 'End ARM']]
ch.columns = ['SR', 'BA', 'EA']
ch.loc[:, 'bs'] = ch.loc[:, 'BA'].apply(roundme)
ch.loc[:, 'es'] = ch.loc[:, 'EA'].apply(roundme)
ch.loc[:, 'UID'] = ch.apply(getUID, axis = 1)

ch_wid = ch[~np.isnan(ch.UID)]
print len(ch_wid)
ch2 = ch[np.isnan(ch.UID)].sort(columns = ['SR', 'BA', 'EA'])
print len(ch2)

ch_ids = list(np.unique(np.array(ch_wid.UID)))
print len(ch_ids)

def fillUID(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    begin = row['bs']
    if (end - bg < 0.15):
        res = db[((db.SR==rd) & (db.bs == begin))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in ch_ids: #  change su_ids to a list and just append x to it if not found in ids.
                ch_ids.append(x)
                return x
            return None
    else:
        return None
def fillUID2(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    begin = row['bs']
    if (end - bg > 0.15):
        res = db[((db.SR==rd) & (db.bs == int((bg*10 + end*10)/2)))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in ch_ids:
                ch_ids.append(x)
                return x
            return None
    else:
        return None
def fillUID3(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    endARM = row['es']
    if (end - bg > 0.05) & (end - bg < 0.5):
        res = db[((db.SR==rd) & (db.es == endARM))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in ch_ids:
                ch_ids.append(x)
                return x
            return None
    return None
def fillUID4(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
#     endARM = row['es']
    if (end - bg > 0.25):
        res = db[((db.SR==rd) & (db.bs == int((bg*10 + end*10)/3)))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in ch_ids:
                ch_ids.append(x)
                return x
            return None
    return None
def fillUID5(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
#     endARM = row['es']
    if (end - bg > 0.25):
        res = db[((db.SR==rd) & (db.es == int((bg*10 + end*10)/3)))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in ch_ids:
                ch_ids.append(x)
                return x
            return None
    return None

ch2.loc[:, 'UID'] = ch2.apply(fillUID, axis = 1)
ch_w2 = ch2[~np.isnan(ch2.UID)]
print len(ch_w2)
ch3 = ch2[np.isnan(ch2.UID)]
print len(ch3)

ch3.loc[:, 'UID'] = ch3.apply(fillUID2, axis = 1)
ch_w3 = ch3[~np.isnan(ch3.UID)]
print len(ch_w3)
ch4 = ch3[np.isnan(ch3.UID)]
print len(ch4)

ch4.loc[:, 'UID'] = ch4.apply(fillUID3, axis = 1)
ch_w4 = ch4[~np.isnan(ch4.UID)]
print len(ch_w4)
ch5 = ch4[np.isnan(ch4.UID)]
print len(ch5)

ch5.loc[:, 'UID'] = ch5.apply(fillUID4, axis = 1)
ch_w5 = ch5[~np.isnan(ch5.UID)]
print len(ch_w5)
ch6 = ch5[np.isnan(ch5.UID)]
print len(ch6)

ch6.loc[:, 'UID'] = ch6.apply(fillUID5, axis = 1)
ch_w6 = ch6[~np.isnan(ch6.UID)]
print len(ch_w6)
ch7 = ch6[np.isnan(ch6.UID)]
print len(ch7)

print len(np.unique(ch_ids))
print len(db)
print len(ch)

ch_all = pd.concat([ch_wid, ch_w2, ch_w3, ch_w4, ch_w5, ch_w6])
ch_all = ch_all.drop_duplicates(subset = ['UID'])

CH[['SR', 'Begin ARM', 'End ARM', u'Major Rehab Year']].head()

ch_final = pd.merge(ch_all, CH, how = 'left', left_on = ['SR', 'BA', 'EA'], right_on = ['SR', 'Begin ARM', 'End ARM'])
ch_FINAL = ch_final.dropna(subset = [u'Major Rehab Year'])
# CDS = cds_FINAL.drop_duplicates(subset = ['SR', 'Begin ARM', 'End ARM', 'Construction End Date'])

ch_FINAL[['SR', 'Begin ARM', 'End ARM','UID', u'Major Rehab Year']].head()

ch_temp = CH.drop_duplicates(subset=['SR', 'Begin ARM', 'End ARM', 'Major Rehab Year'])
print len(ch_temp)

print len(ch_FINAL)
print len(ch_ids)
print len(ch)

ch_FINAL.to_csv('../FinalDatasets/ContractsHistory.csv')

ch_sorted = ch_FINAL.sort(columns = ['SR', 'Begin ARM', 'End ARM'])

ch_sorted.head()

su_final.head()

cds_all.head()

ch_all.head()

scc = pd.merge(su_final, cds_all, how = 'left', left_on = ['UID'], right_on = ['UID'])

scc2 = scc.dropna(subset = ['SR_y'])
print len(scc), len(scc2)

scc3 = pd.merge(scc2, ch_all, how = 'left', left_on = ['UID'], right_on = ['UID'])
scc3.head()

sccT = scc3.dropna(subset = ['SR'])
print len(sccT), len(scc3), len(scc2)
