==> 02. Merging Datasets for creating crack-initiation model dataset.ipynb <==
import numpy as np
import pandas as pd
import xlsxwriter

dbT= pd.read_csv("../tables/Traffic to Contract History.txt", header = None)
exdbT= pd.read_excel("../DataFiles/Traffic to Contract History.xlsx")
dbCH= pd.read_csv("../FinalDatasets/ContractsHistory.csv")
dbCO= pd.read_csv("../FinalDatasets/Contracts.csv")
SU = pd.read_csv("../FinalDatasets/SurveyUnits.csv")
SDD = pd.read_csv("../tables/Survey Distress Details.txt", header = None)
exSDD = pd.read_excel("../DataFiles/Survey Distress Details.xlsx")
# SDD = pd.read_excel("../DataFiles/PSCdata.xlsx")

dbT.columns = exdbT.columns
SDD.columns = exSDD.columns

#dbT = dbT.drop(['Single Trucks', 'Double Trucks', 'Peak Hour Percentage', 'Peak Hour Directional Split'], axis =1)
dbT = dbT[['Traffic Year', 'Traffic Year ESALs _One Direction_', 'Contract History ID']]
dbT.columns = ['Traffic Year', 'ESALs', 'Contract History ID']

dbCH = dbCH[['SR','Begin ARM', 'End ARM', 'Region', 'ID', 'Construction Year', 'Major Rehab Year', 'Segment Lane Miles', 
            'ESALs Since Major Rehab', 'Total Thickness', 'Total Surface Thickness', 'Listed Surface Type', 'Pavement Type',
            'Construction Pavement Type', 'Major Rehab Type', 'Prior Major Rehab Year', 'Prior Major Rehab Type', 'UID']]

dbT.head()

# Merging Contracts History Dataset to Traffic Loading (Y2013)

print len(dbT)
print len(np.unique(np.array(dbT.loc[:, 'Contract History ID'])))
print len(dbCH)

T_CH = pd.merge(dbT, dbCH, how = 'left',left_on = ['Contract History ID'], right_on = ['ID'])
tch = T_CH.dropna(subset =['UID'])

print len(T_CH)
print len(dbT) == len(T_CH) # If false=OK, as there might be more than one contract for a given segment.
print len(tch)

# Contracts Database

dbC = dbCO[['SR', 'Begin ARM', 'End ARM', 'ID','Construction End Date','Substantial Completion Date','Contract Type Description', 'Surface Type Code', 'Surface Type Description', 'Surface Thickness',
          'Treat Base Type Code', 'Treat Base Type Description', 'Treat Base Thickness', 'UTreat Base Thickness', 'UID']]

# Function extracts the year value from the date string.
def getYear(year):
    return int(float(year.split('/')[2].split(' ')[0]))

# Function returns year values from a column with date format.
def returnYear(row):
    date = row['Construction End Date']
    year = getYear(date)
    return year

# Construction end year: Year that the specified project was ended.
dbC.loc[:, 'cYear'] = dbC.apply(returnYear, axis = 1)

dbC[['SR', 'Begin ARM', 'End ARM', 'UID', 'Construction End Date', 'cYear']].head()

I have split up the construction date column and created a 'cYear' column which I will use to select the actions which were performed before year 2014. 

dbC2 = dbC[dbC.loc[:, 'cYear'] <= 2013]

print len(dbC)
print len(dbC2)

dbCs = dbC2

len(dbCs)

print len(dbCs[dbCs.loc[:, 'UTreat Base Thickness']> 0.0])
print len(dbCs[dbCs.loc[:, 'Treat Base Thickness']> 0.0])
print len(dbCs)

From the above summary, it can be observed that the number of segments where the treated or non-treated base layer thickness was recorded is less than 0.2 percent of the total segments in the dataset. The base thickness variable will therefore be removed for the models using records of year less than 2000. It should be noted, however, that for records of year >= 2000, there is a significant number of records with treated and untreated base-layer thicknesses so included of these variables into the models can be warranted.

# db = dbCs[['SR', 'Begin ARM', 'End ARM','uid', 'ID','cYear','Contract Type Description', 'Surface Type Code', 'Surface Type Description', 'Surface Thickness']]
db = dbCs #leaving all the variables in the dataset

print len(dbCH), len(db)

# Merging Contracts and merged Traffic-Contract History dataset.

print len(db)
print len(tch)

tch.columns

C_CH = pd.merge(db, tch, how = 'left',left_on = ['UID'], right_on = ['UID'])
CCHs = C_CH.dropna(subset = [u'Major Rehab Year'])

print len(CCHs), len(C_CH)

pvmtType = np.unique(np.array(CCHs[['Pavement Type']]))
print pvmtType

# selected pavement type
spt = [pvmtType[1], pvmtType[2], pvmtType[3], pvmtType[4],  pvmtType[6]]
print spt

CCHasp = CCHs[CCHs.loc[:,'Pavement Type'].isin(spt)] #selecting only asphalt pavements

Extracting the Surface types used for rehabilitation

CCHasp.index = np.arange(len(CCHasp))

MatType = np.unique(np.array(CCHasp.loc[:, 'Surface Type Code']))

MTdic = {}
for mt in MatType[1:]:
    MTdic[mt] = len(CCHasp[CCHasp.loc[:, 'Surface Type Code']==mt])

uniqMT = []
for mt in MTdic.keys():
    if MTdic[mt] > 850:
        print mt, MTdic[mt]
        uniqMT.append(mt)

uniqMT

# # CCHs[~CCHs.loc[:,'Surface Type Code'].isin(uniqMT)].loc[:,'Surface Type Code']
# for i in CCHs.index:
#     if CCHs.loc[i,'Surface Type Code'] not in uniqMT:
#         CCHs.loc[i,'Surface Type Code'] = 'Other'  

I had categorized the surface type previously but later realized that the final dataset might not contain the same proportion of variables, given that I am going to consider only asphalt pavements. Categorization of this nature will be among the final steps of the data cleaning process..

Just concluded categorizing the Surface Type Code accordingly. Next step is to link the CCHs (Contract and Contract-History) dataframe to the Survey_Units dataframe using the linker dataframe uploaded below..

len(CCHasp[CCHasp.cYear == CCHasp.loc[:, 'Major Rehab Year']])

# Merging survey units (SU) and Contracts and Contract History Datasets (CCHa)

# SU.columns

su = SU[[u'SR', u'Begin ARM', u'End ARM', 'UID', u'Is Bridge', u'Lane Surface Type', u'Lane Miles', u'Region', 
         u'Federal Functional Class', u'Int ID']]

su.head()

print len(CCHasp), len(su)

SCCH = pd.merge(su, CCHasp, how = 'left',left_on = ['UID'], right_on = ['UID'])
scch = SCCH.dropna(axis=0, how='any', thresh=None, subset= ['cYear'], inplace=False)

print len(scch)

# Merging joint dataset with survey distress details

SDD.head()

SD = pd.read_csv("../tables/Survey Details.txt", header = None)
exSD = pd.read_excel("../DataFiles/SurveyDetails.xlsx")
SD.columns = exSD.columns

sd = SD[[u'Year','Rutting', 'IRI','PSC', u'Int ID']]
sd.columns = [u'sYear','Rutting', 'IRI','PSC', u'IntID']

SDDs = SDD[[                                 u'Year',
                      u'Low Alligator Cracking',
                   u'Medium Alligator Cracking',
                     u'High Alligator Cracking',
          u'Low Flexible Longitudinal Cracking',
       u'Medium Flexible Longitudinal Cracking',
         u'High Flexible Longitudinal Cracking',
                       u'Low Flexible Patching',
                    u'Medium Flexible Patching',
                      u'High Flexible Patching',
            u'Low Flexible Transverse Cracking',
         u'Medium Flexible Transverse Cracking',
           u'High Flexible Transverse Cracking',
                                      u'Int ID']]

SDDs.columns = [                                 u'sYear',
                      u'Low Alligator Cracking',
                   u'Medium Alligator Cracking',
                     u'High Alligator Cracking',
          u'Low Flexible Longitudinal Cracking',
       u'Medium Flexible Longitudinal Cracking',
         u'High Flexible Longitudinal Cracking',
                       u'Low Flexible Patching',
                    u'Medium Flexible Patching',
                      u'High Flexible Patching',
            u'Low Flexible Transverse Cracking',
         u'Medium Flexible Transverse Cracking',
           u'High Flexible Transverse Cracking',
                                      u'IntID']

print len(SDDs)
print len(sd)

The following code merges the Survey Details and Survey Distress Details datasets in order to identify the years where there was measurement for the respective distresses, especially for alligator and longitudinal cracking

SDDs.loc[:, 'temp'] = 28

SDDm = pd.merge(sd, SDDs, how = 'left',left_on = ['sYear', 'IntID'], right_on = ['sYear', 'IntID'])

sdd = SDDm.dropna(subset = ['temp', 'PSC'])
print len(sdd)
print len(SDDm)

sdd = sdd.fillna(0)

sdd.loc[:, 'Alli'] = np.array(sdd.loc[:, 'Low Alligator Cracking']) + np.array(sdd.loc[:, 'Medium Alligator Cracking']
                    ) + np.array(sdd.loc[:, 'High Alligator Cracking'])
sdd.loc[:, 'Long'] = np.array(sdd.loc[:, 'Low Flexible Longitudinal Cracking']) + np.array(sdd.loc[:, 'Medium Flexible Longitudinal Cracking']
                    ) + np.array(sdd.loc[:, 'High Flexible Longitudinal Cracking'])

sdd.head()

Code below merges the two datasets, total and sdd.. (scct, and survey distress details)

allData = pd.merge(sdd, scch, how = 'left',left_on = ['IntID'], right_on = ['Int ID'])
cdata = allData.dropna(axis=0, how='any', thresh=None, subset= [u'UID'], inplace=False)

print len(allData)
print len(cdata), len(sdd)
print len(cdata) == len(sdd)
print len(np.unique(np.array(cdata.UID)))
print len(cdata[cdata.loc[:,'sYear']>1980])
print len(np.unique(np.array(cdata[cdata.loc[:,'sYear']>1980].loc[:, 'UID'])))

len(np.unique(np.array(cdata.loc[:, [u'SR', u'Begin ARM', u'End ARM']])))

CDATA = cdata[[ u'sYear', u'Rutting', u'IRI', u'PSC', u'IntID', 'Alli', 'Long', u'Low Flexible Patching',
                    u'Medium Flexible Patching', u'High Flexible Patching',
            u'Low Flexible Transverse Cracking',
         u'Medium Flexible Transverse Cracking',
           u'High Flexible Transverse Cracking',
                                          u'SR',
                                   u'Begin ARM',
                                     u'End ARM',
                                         u'UID',
                                   u'Is Bridge',
                           u'Lane Surface Type',
                                  u'Lane Miles',
                                    u'Region_x',
                    u'Federal Functional Class',
                       u'Construction End Date',
                 u'Substantial Completion Date',
                   u'Contract Type Description',
                           u'Surface Type Code',
                    u'Surface Type Description',
                           u'Surface Thickness',
                        u'Treat Base Type Code',
                 u'Treat Base Type Description',
                        u'Treat Base Thickness',
                       u'UTreat Base Thickness',
                                       u'cYear',
                                u'Traffic Year',
                                       u'ESALs',
                         u'Contract History ID',
                           u'Construction Year',
                            u'Major Rehab Year',
                          u'Segment Lane Miles',
                     u'ESALs Since Major Rehab',
                             u'Total Thickness',
                     u'Total Surface Thickness',
                         u'Listed Surface Type',
                               u'Pavement Type',
                  u'Construction Pavement Type',
                            u'Major Rehab Type',
                      u'Prior Major Rehab Year',
                      u'Prior Major Rehab Type']]

def change2Percentage(row):
    p = row[u'Alli']
    m = row[u'Lane Miles']
    if not (p == 'nan'):
        return p/(m*5280) * 100
    else:
        return 'nan'  

def change2PercentageL(row):
    p = row[u'Long']
    m = row[u'Lane Miles']
    if not (p == 'nan'):
        return p/(m*5280) * 100
    else:
        return 'nan' 

CDATA.loc[:, 'AlliP'] = CDATA.apply(change2Percentage, axis = 1)
CDATA.loc[:, 'LongP'] = CDATA.apply(change2PercentageL, axis = 1)

MatType = np.unique(np.array(ds.loc[:, 'Surface Type Code']))
tds = CDATA.drop_duplicates(subset = ['UID'])

MTdic = {}
for mt in MatType[1:]:
    n = len(tds[tds.loc[:, 'Surface Type Code']==mt])
    if n > 500:
        MTdic[mt] = n
print len(MTdic)

MTdic

print len(tds)

def catSurfType(row):
    st = row['Surface Type Code']
    if st in MTdic.keys():
        return st
    else:
        return 'Other'  

CDATA.loc[:, "STcode"] = CDATA.apply(catSurfType, axis = 1)

CDATA.to_csv('../FinalDatasets/workingDataset.csv')
