==> 03. Checking model sample size validity.ipynb <==
import pandas as pd
import numpy as np
import xlsxwriter

ds = pd.read_csv('../FinalDatasets/workingDataset.csv')

ds.head()

ds.columns

np.unique(np.array(ds.loc[:,'Pavement Type']))

rds = ds[ds.cYear == ds.loc[:,'Major Rehab Year']]
print len(rds)

df = ds

# cids = df[df.AlliP < 5]
# print len(cids)
# cd = cids.sort(columns = ['uid'])
# cd.head(20)

segments = []
for i in np.unique(np.array(df.UID)):
    sdf = df[df.UID == i]
    years = np.array(sdf.sYear)
    rehab = np.array(sdf.loc[:, 'Major Rehab Year'])[0]
    if rehab in years:
        if (rehab -1) in years:
            if (rehab+1) in years:
                if (rehab+2) in years:
                    segments.append(i)

len(segments)

segmentsCPM = []
for i in np.unique(np.array(df.UID)):
    sdf = df[df.UID == i]
    years = np.array(sdf.sYear)
    rehab = np.array(sdf.loc[:, 'Major Rehab Year'])[0]
    if rehab in years:
            if (rehab+1) in years:
                if (rehab+2) in years:
                    if (rehab+3) in years:
                        if (rehab+4) in years:
                            segmentsCPM.append(i)

print len(segmentsCPM)

segCPM = []
for i in np.unique(np.array(df.UID)):
    sdf = df[df.UID == i]
    years = np.array(sdf[sdf.sYear < 2000].sYear)
    if len(years) > 5:
        segCPM.append(i)

print len(segCPM)

cpmDataset = df[df.UID.isin(segCPM)]

print len(cpmDataset)
print len(np.unique(np.array(cpmDataset.UID)))

cpm = cpmDataset[[u'sYear', u'Rutting', u'IRI', u'PSC', u'IntID', u'Alli',
       u'Long', u'Low Flexible Patching', u'Medium Flexible Patching',
       u'High Flexible Patching', u'Low Flexible Transverse Cracking',
       u'Medium Flexible Transverse Cracking',
       u'High Flexible Transverse Cracking', u'SR', u'Begin ARM', u'End ARM',
       u'UID', u'Is Bridge', u'Lane Surface Type', u'Lane Miles', u'Region_x',
       u'Federal Functional Class', u'Construction End Date',
       u'Substantial Completion Date', u'Contract Type Description',
       u'Surface Type Code', u'Surface Type Description', u'Surface Thickness',
       u'Treat Base Type Code', u'Treat Base Type Description',
       u'Treat Base Thickness', u'UTreat Base Thickness', u'cYear',
       u'Traffic Year', u'ESALs', u'Contract History ID', u'Construction Year',
       u'Major Rehab Year', u'Segment Lane Miles', u'ESALs Since Major Rehab',
       u'Total Thickness', u'Total Surface Thickness', u'Listed Surface Type',
       u'Pavement Type', u'Construction Pavement Type', u'Major Rehab Type',
       u'Prior Major Rehab Year', u'Prior Major Rehab Type', u'AlliP',
       u'LongP', u'STcode']]

cpm.to_csv('../FinalDatasets/CPMdataset.csv')

cpm.corr()

cpm[['sYear', 'UID', 'IRI', 'AlliP', 'Rutting']].head(20)

segDic = {}
for i in np.unique(np.array(idf.uid)):
    segDic[i] = []
    sdf = idf[idf.uid == i]
    years = np.array(sdf.sYear)
    rehab = np.array(sdf.loc[:, 'Major Rehab Year'])[0]
    segDic[i].append((rehab))
    if rehab in years:
        if (rehab -1) in years:
            segDic[i].append((rehab-1))
            if (rehab+1) in years:
                segDic[i].append((rehab+1))
                if (rehab+2) in years:
                    segDic[i].append((rehab+2))
                    if (rehab+3) in years:
                        segDic[i].append((rehab+3))

segDic

mIDs = []
for i in idf.index:
    if idf.loc[i, 'sYear'] in segDic[idf.loc[i, 'uid']]:
        mIDs.append(i)

print len(mIDs)

mdata = idf[idf.index.isin(mIDs)]

print len(mdata) == len(mIDs)

mdata = mdata.sort(columns = ['uid'])

pdata = mdata[~np.isnan(mdata.AlliP)]

print len(np.unique(np.array(pdata.uid))), len(pdata)

mdata.columns
