==> 04. Adding Coordinates and Traffic Data (ESALs) to Crack Progression Model Dataset.ipynb <==
import pandas as pd
import numpy as np

ds = pd.read_csv("../FinalDatasets/CPMdataset.csv")

wd = pd.read_excel('../DataFiles/AnnualWeatherR2.xlsx')
wds = pd.read_excel('../DataFiles/AnnualWeatherR.xlsx')
wds = wds[['Year', 'FTCycles', 'Airport']]
wd = pd.concat([wds, wd], axis = 0)

wd.head()

def getCode(row):
    index = row['IntID']
    return ld.loc[index,'TargetID']

# Adding traffic data

dc = ds.drop_duplicates(subset = [u'UID'])
df = dc[['UID', 'Federal Functional Class', 'ESALs']]

df.columns = ['UID', 'Functional_Class', '2013']

def getClass(row):
    f = row['Functional_Class']
    req = f.split()[0]
    return req

df.loc[:,'Class'] = df.apply(getClass, axis = 1)

df.head()

de = pd.read_excel("../DataFiles/EsalIncrements.xlsx") 

de.head()

di = de[['Year', 'pRural', 'pUrban']]
di.columns = ['Year', 'Rural', 'Urban']

def extrapolate(df):   
    for row in df.index:
        if df.loc[row, 'Class'] == 'Rural':
            cat = 'r'
        else:
            cat = 'u'
        for year in (np.arange(32)+1981)[::-1]:
            name = str(year)
            pname = str(year-1)
            ind  = year + 1 - 1980
            df.loc[row, name] = (df.loc[row, pname])/(1 + di.loc[ind, cat])

di.index = np.array(di.Year)

di.head()

def extrapolate(row, year): 
    funClass = row['Class']
#     year  = 2013
    esal  = row[str(year)]
    r     = di.loc[year, funClass]
    required = esal/(1+r)
    return required

# extrapolate(df)
for year in (np.arange(32)+1981)[::-1]:
    name = str(year)
    pYear = year+1
    df.loc[:, name] = df.apply(extrapolate, axis =1, args = (pYear,))

(np.arange(32)+1981)[::-1]

df.head()

df2 = df[[u'2013',u'2012', u'2011', u'2010', u'2009', u'2008', u'2007', u'2006', u'2005', u'2004', u'2003', u'2002',
          u'2001', u'2000', u'1999',u'1998', u'1997', u'1996', u'1995', u'1994', u'1993', u'1992', u'1991',  u'1990',
          u'1989', u'1988', u'1987',u'1986', u'1985', u'1984', u'1983',u'1982', u'1981']]

df2.index = np.array(df.loc[:, 'UID'])

df2.head()

df2_reordered = df2[df2.columns[::-1]]
df2_reordered.head()

df2_cum_sum = df2_reordered.cumsum(axis=1)
df2_cum_sum.head()

df2_cum_sum['id'] = df2_cum_sum.index

desired_table = df2_cum_sum.stack()
desired_table.head()

desired_table.head(50)

def fillESALs(row):
    index = row['UID']
    value = str(row['sYear'])
    return df2.loc[index, value]

sds = ds[ds.sYear < 2014]

#"Extrapolated" ESALs..
sds.loc[:,'ESALs'] = sds.apply(fillESALs, axis = 1)

#Exporting the almost complete dataset
sds.to_csv("../FinalDatasets/CPMdatasetWesals.csv") 

sds.head()

# Adding coordinates to dataset to identify segment locations

milePost = pd.read_csv("../FinalDatasets/mileposts.csv")

milePost.head()

routeNames = np.unique(np.array(milePost.SR))
print routeNames

# for name in routeNames:
#     if len(name) > 3:
#         print int(float(re.findall("[-+]?\d+[\.]?\d*", name)[0]))

import re
def change2ints(row):
    sr = row['SR']
    if len(sr) > 3:
        return int(float(re.findall("[-+]?\d+[\.]?\d*", name)[0]))
    return int(sr)

milePost.loc[:,'SR'] = milePost.apply(change2ints, axis = 1)

db = milePost[['LONGITUDE', 'LATITUDE', 'MP_VALUE', 'SR']] 

db.head()

print int(1.98)
sr = 2
bmp = 0.99
print np.array(db[((db.loc[:,'MP_VALUE'] > bmp) & (db.loc[:,'MP_VALUE'] < (bmp+2))) & (db.loc[:,'SR'] == sr)])

def getLat(row):
    sr = row['SR']
    bmp = int(row['Begin ARM'])
#     emp = int(row['End ARM'])
    res = np.array(db[((db.loc[:,'MP_VALUE'] > (bmp-1)) & (db.loc[:,'MP_VALUE'] < (bmp+4))) & (db.loc[:,'SR'] == sr)])
    if len(res) > 0:
        coordinates = res[0][0:2]
        x = coordinates[0]
        return coordinates[1]
    return None
sds.loc[:,'Latitude'] = sds.apply(getLat, axis = 1)

def getLong(row):
    sr = row['SR']
    bmp = int(row['Begin ARM'])+1
#     emp = int(row['End ARM'])+1
    res = np.array(db[((db.loc[:,'MP_VALUE'] > (bmp-1)) & (db.loc[:,'MP_VALUE'] < (bmp+4))) & (db.loc[:,'SR'] == sr)])
    if len(res) > 0:
        coordinates = res[0][0:2]
        return coordinates[0]
#         return coordinates[1]
    return None
sds.loc[:,'Longitude'] = sds.apply(getLong, axis = 1)

sds.head()

cdf = sds.drop_duplicates(subset = ['UID'])

cleaned = cdf[~np.isnan(cdf.Longitude)]
cleaned = cleaned[~np.isnan(cleaned.Latitude)]

print len(cdf)
print len(cleaned)

cleaned.index = range(len(cleaned))
coordata = cleaned[['UID', 'Longitude', 'Latitude']]

coordata.to_csv('../FinalDatasets/RCsegCoord.csv')

sds = sds[~np.isnan(sds.Longitude)]
sds = sds[~np.isnan(sds.Latitude)]

print len(sds)

sds.to_csv('../FinalDatasets/RoughnessCrackingDataset.csv')

sds[['UID', 'cYear', 'Major Rehab Year', 'Prior Major Rehab Year']].head()

print len(np.unique(np.array(sds[sds.cYear > 1970].UID)))

df.head()

cumdf = df[['UID', u'1981']]

segments = np.unique(np.array(df.UID))
msegs = np.tile(segments, 33)
print len(msegs) == len(segments) *33

years = np.arange(33)+1981
myears = np.tile(years, len(segments))
print len(msegs) == len(myears)

sorted_segments = np.sort(msegs)
print sorted_segments[:40]
print msegs[0:40]
print myears[:40]

segDic = {'ID': sorted_segments, 'Year': myears}

cumdf = pd.DataFrame(segDic)

cumdf.head()

print len(cumdf)

def getEsal(row, year):
    id = row['ID']
    name = str(year)
    return df2.loc[id, name]

for year in years:
    cumdf.loc[:, str(year)] =cumdf.apply(getEsal, axis =1, args = (year,))
