==> 05. RasterFile Data Abstraction copy.ipynb <==
import pandas as pd
import numpy as np
import gdal
from gdalconst import * 
import os, sys, time
import xlsxwriter

def getValues(rasterDS, locData, rasterName):
    size = len(locData)
    startTime = time.time() #if timing the function speed is really necessary
    xValues = np.array(locData.loc[:, 'Longitude'])
    yValues = np.array(locData.loc[:, 'Latitude'])
    rows = rasterDS.RasterYSize #number of rows
    cols = rasterDS.RasterXSize # number of columns
    bands = rasterDS.RasterCount # number of data value types (it's going to be 1 for our case)
    
    # get georeference info
    transform = rasterDS.GetGeoTransform()
    xOrigin = transform[0]
    yOrigin = transform[3]
    pixelWidth = transform[1]
    pixelHeight = transform[5]
    
    band = rasterDS.GetRasterBand(1)
    # getting all data at once
    allData = band.ReadAsArray(0,0, cols, rows)
    
    #Result dataset
    reqDataset = locData
    reqDataset[rasterName] = 'nan'
    
    for i in range(size):
        # get x, y
        x = xValues[i]
        y = yValues[i]
        # compute pixel offset
        xOffset = int((x - xOrigin) / pixelWidth)
        yOffset = int((y - yOrigin) / pixelHeight)
        value   = allData[yOffset, xOffset]
        reqDataset.loc[i, rasterName] = value
    endTime = time.time()
    print 'The script took ' + str(endTime - startTime) + 'seconds'
    
    return reqDataset

# baseFile = 'C:\\Users\\Allan\\Dropbox\\2015\\Research Data\\Weather Variables\\'
# coordDataset = dataset
# link = '\\RasterFiles\\'
def getDataset(RasterID, cat, coordDataset, baseFile, link):
    rasterName = str(RasterID) # RasterID might be either a string(month) or an int (year)
    fileName = baseFile + cat + link + rasterName + '.asc'
    ds = gdal.Open(fileName)
    if ds is None:
        print 'could not open ' + fileName
        sys.exit(1)
    completeDataset = getValues(ds, coordDataset, rasterName)
#     return completeDataset

def RasterFilesData2Excel(variable, coordDataFrame):
    baseFile = '../RasterFiles/'
    resFolder = '../Weather Variables/'
    cal      = {1:'jan', 2: 'feb', 3:'mar', 4:'apr', 5:'may', 6:'jun', 7:'jul', 8:'aug', 9: 'sep',
               10:'oct', 11:'nov', 12:'dec'}
    if variable == 'Precipitation':
        link = '/'
        for year in np.arange(33)+1981:
            getDataset(year, variable, coordDataFrame, baseFile, link)
        fileName = resFolder+ 'Ppt/' + 'PrecipitationRecord1981_2013.xlsx'
        writer = pd.ExcelWriter(fileName)
        coordDataFrame.to_excel(writer)
    elif variable == 'MinTemp':
        fileName = resFolder + variable + '/'
        for year in np.arange(33)+1981:
            link = '/' + str(year)+'/'
            record = fileName + str(year) + '.xlsx'
            for month in np.arange(12)+1:
                getDataset(cal[month], variable, coordDataFrame, baseFile, link)
            writer = pd.ExcelWriter(record)
            coordDataFrame.to_excel(writer, sheet_name=str(year))
            coordDataFrame = dataset[['UID', 'Longitude', 'Latitude']]
    elif variable == 'MaxTemp':
        fileName = resFolder + variable + '/'
        for year in np.arange(33)+1981:
            link = '/' + str(year)+'/' 
            record = fileName + str(year) + '.xlsx'
            for month in np.arange(12)+1:
                getDataset(cal[month], variable, coordDataFrame, baseFile, link)
            writer = pd.ExcelWriter(record)
            coordDataFrame.to_excel(writer, sheet_name=str(year))
            coordDataFrame = dataset[['UID', 'Longitude', 'Latitude']]
    else:
        print 'Incorrent variable name. The accepted variable names include "Precipitation", "MinTemp", and "MaxTemp"' 
        
    

dataFile = '../FinalDatasets/RCsegCoord.csv'
dataset = pd.read_csv(dataFile)

dataset = dataset[~np.isnan(dataset.Latitude)]
print len(dataset)

PrecipitationDS = dataset[['UID', 'Longitude', 'Latitude']]
MinTemperatureDS = dataset[['UID', 'Longitude', 'Latitude']]
MaxTemperatureDS = dataset[['UID', 'Longitude', 'Latitude']]

RasterFilesData2Excel('MinTemp', MinTemperatureDS)

RasterFilesData2Excel('MaxTemp', MaxTemperatureDS)

RasterFilesData2Excel('Precipitation', PrecipitationDS)
