==> 06. Merging weather variables to almost complete dataset for Roughness Model.ipynb <==
import pandas as pd
import numpy as np
import xlsxwriter

# Min temp data basefile
mnfile = "../Weather Variables/MinTemp/" 
# Max temp data basefile
mxfile = "../Weather Variables/MaxTemp/" 
# Precipiation data basefile
pfile = "../Weather Variables/Ppt/" 

def findMaxMin(db, code):
    if code == 'max':
        return db.max(axis=1)
    else:
        return db.min(axis=1)

def summarizeWeatherData(var):
    if var == 'maxT':
        bfile = mxfile
        code = 'max'
    elif var == "minT":
        bfile = mnfile
        code = 'min'
    else:
        bfile = pfile
        return pd.read_excel(bfile+'PrecipitationRecord1981_2013.xlsx')

    tdb = pd.read_excel(bfile+str(1991)+'.xlsx')
    merged = tdb[['UID']]
    for year in np.arange(33)+1981:
        db = pd.read_excel(bfile+str(year)+'.xlsx')
        adb = db[[u'jan',  u'feb',  u'mar',  u'apr', u'may',  u'jun',  u'jul',  u'aug',  u'sep',  u'oct',  u'nov',  
                  u'dec']]
        req = findMaxMin(adb, code)
#         print np.array(req)[0:3]
        req.name = str(year)
        merged = merged.join(req)
    return merged

minTemp = summarizeWeatherData('minT')

maxTemp = summarizeWeatherData('maxT')

Ppt = summarizeWeatherData('ppt')

Ppt.head()

minTemp.index = np.array(minTemp.UID)
maxTemp.index = np.array(maxTemp.UID)
Ppt.index = np.array(Ppt.UID)

segments = np.unique(np.array(Ppt.UID))
msegs = np.tile(segments, 33)
sorted_segments = np.sort(msegs)
print len(sorted_segments) == len(segments) *33
years = np.arange(33)+1981
myears = np.tile(years, len(segments))
print len(sorted_segments) == len(myears)

segDic = {'ID': sorted_segments, 'Year': myears}
cumdf = pd.DataFrame(segDic)

print len(cumdf)

def createCumDF(cumdf, df): 
    tcumdf = cumdf.copy()
    for x in segments:
        for year in years:
            tcumdf[tcumdf.ID == x].loc[:, str(year)] = df.loc[x, str(year)]
    return tcumdf

cumPpt = createCumDF(cumdf, Ppt)

df = pd.read_excel('../modelData/cdata_cim.xlsx')

len(df[df.sYear < 2014]) == len(df)

def popPPT(row):
    index = row['uid']
    year  = row['sYear']
    val   = str(year)
    return Ppt.loc[index, val]
df.loc[:, 'PPT'] = df.apply(popPPT, axis = 1)

def popMinT(row):
    index = row['uid']
    year  = row['sYear']
    val   = str(year)
    return minTemp.loc[index, val]
df.loc[:, 'MinTemp'] = df.apply(popMinT, axis = 1)

def popMaxT(row):
    index = row['uid']
    year  = row['sYear']
    val   = str(year)
    return maxTemp.loc[index, val]
df.loc[:, 'MaxTemp'] = df.apply(popMaxT, axis = 1)

df.head()

df.to_excel('../modelData/CIMdataset2.xlsx')
