==> 13. Adding weather variables to modefied roughness model dataset.ipynb <==
import pandas as pd
import numpy as np

ppt = pd.read_csv('../FinalDatasets/ppt.csv')
maxt = pd.read_csv('../FinalDatasets/MaxT.csv')
mint = pd.read_csv('../FinalDatasets/MinT.csv')
esals =pd.read_csv('../FinalDatasets/ESALs.csv')
rdat = pd.read_csv('../modelDatasets/rmdata1988_1999.csv')

print len(np.unique(np.array(esals.ID)))*33

selected_ids = np.unique(np.array(ppt.ID))
esals = esals[esals.ID.isin(selected_ids)]

esals = esals.drop_duplicates(subset = ['ID'])
ppt  = ppt.drop_duplicates(subset = ['ID'])
mint = mint.drop_duplicates(subset = ['ID'])
maxt = maxt.drop_duplicates(subset = ['ID'])

print len(esals)
print len(ppt)
print len(mint)
print len(maxt)
print len(rdat)

esals.index = esals.ID
ppt.index = ppt.ID
mint.index = mint.ID
maxt.index = maxt.ID

esals.head()

mint.head()

# esals.to_csv('../modelDatasets/esals.csv')
# ppt.to_csv('../modelDatasets/ppt.csv')
# mint.to_csv('../modelDatasets/mint.csv')
# maxt.to_csv('../modelDatasets/maxt.csv')

esals.columns

rdat.head()

print len(rdat[rdat.UID.isin(mint.ID)]) == len(rdat)

rdat.head()

print rdat.UID.dtype
print ppt.ID.dtype

def addVariable(row, df):
    index = int(row['UID'])
    year  = str(row['sYear'])
    return df.loc[index, year]
def addPPT(row):
    index = int(row['UID'])
    year  = str(row['sYear'])
    return ppt.loc[index, year]
def addMinT(row):
    index = int(row['UID'])
    year  = str(row['sYear'])
    return mint.loc[index, year]
def addMaxT(row):
    index = int(row['UID'])
    year  = str(row['sYear'])
    return maxt.loc[index, year]

rdat.loc[:, 'esals'] = rdat.apply(addVariable, axis = 1, args = (esals,))
rdat.loc[:, 'ppt'] = rdat.apply(addPPT, axis = 1)
rdat.loc[:, 'mint'] = rdat.apply(addMinT, axis = 1)
rdat.loc[:, 'maxt'] = rdat.apply(addMaxT, axis = 1)

# rdat.loc[:, 'ppt'] = rdat.apply(addVariable, axis = 1, args = (ppt,))
# rdat.loc[:, 'maxt'] = rdat.apply(addVariable, axis = 1, args = (maxt,))
# rdat.loc[:, 'mint'] = rdat.apply(addVariable, axis = 1, args = (mint,))

rdat.head()

rdat.to_csv('../modelDatasets/RMdata88_99.csv')

# Adding Freeze Thaw Cycles

cpmSegAp = pd.read_csv('../FinalDatasets/CPM_seg_icao.csv')

cpmSegAp.head()

cpmSegAp.index = cpmSegAp.InputID
cpmSegAp.head()

wds = pd.read_excel('../DataFiles/AnnualWeatherR.xlsx') #contains yrs 1981 -1999
wd = wds[['Year', 'FTCycles', 'Airport']]
ft = wd.pivot('Airport', 'Year', 'FTCycles')

ft.head()

def getCode(row):
    index = row['UID']
    return cpmSegAp.loc[index,'TargetID']

print len(rdat[rdat.UID.isin(cpmSegAp.index)])

# rdat.loc[:, 'ICAO'] = rdat.apply(getCode, axis =1)

# Match the UIDs then add the corresponding freeze that cycles
You might have to save a separate csv file of unique IDs first then use QGIS to do the matching.
