==> 14. Adding Freeze Thaw Cycles and Completing Refined Roughness Model Dataset.ipynb <==
import pandas as pd, numpy as np

rdat = pd.read_csv('../modelDatasets/RMdata88_99.csv')

rdat.head()

uids = rdat.UID.unique()

print len(uids)

cds = pd.read_csv('../FinalDatasets/RCSegCoord.csv')

cds.head()

print len(rdat[rdat.UID.isin(cds.UID)]) == len(rdat)

print len(cds) == len(cds.UID.unique())

cds.UID.dtype

cds.UID = cds.UID.astype('int')

cds.to_csv('../FinalDatasets/RCSegCoord.csv')

mds = pd.read_csv('../FinalDatasets/RMid2icao.csv') # contains match data for UID and ICAO codes to obtain FTCycle data

mds.head()

mds.index = mds.InputID

def getCode(row):
    index = row['UID']
    return mds.loc[index,'TargetID']

rdat.UID = rdat.UID.astype('int')

rdat.loc[:, 'ICAO'] = rdat.apply(getCode, axis =1)

Loading the weather dataset with all the variables. Variable of interest is the freeze thaw cycles..

wds = pd.read_excel('../DataFiles/AnnualWeatherR.xlsx') #contains yrs 1981 -1999
wd = wds[['Year', 'FTCycles', 'Airport']]
ft = wd.pivot('Airport', 'Year', 'FTCycles')

ft.head()

print ft.columns.dtype
print rdat.sYear.dtype

def getFTC(row):
    index = row['ICAO']
    year  = row['sYear']
    return ft.loc[index, year]

rdat.loc[:, 'ftc'] = rdat.apply(getFTC, axis = 1)

rdat.head()

rdat.columns

fmd = rdat[[u'UID', u'sYear', u'Rutting', u'IRI', u'PSC', u'Alli', u'Long', u'Is Bridge', u'Lane Miles',
       u'Segment Lane Miles', u'ESALs Since Major Rehab', u'sfThick',
       u'Tthick', u'TSfThick', u'STcode', u'cumPPT', u'meanMinTemp',
       u'meanMaxTemp', u'cumESALs', u'cumFTC', u'pAlli', u'pLong', u'IRIsi',
       u'esals', u'ppt', u'mint', u'maxt', u'ICAO', u'ftc']]

fmd.to_csv('../modelDatasets/rmData1115.csv')
