==> 15. Recreating roughness model dataset.ipynb <==
import pandas as pd, numpy as np

ds = pd.read_csv('../modelDatasets/ndata.csv')

ds.head()

df = pd.read_csv('../tables/Traffic.txt', header=None)
exdf = pd.read_excel("../DataFiles/Traffic.xlsx")

df.columns = exdf.columns

print len(df)
df.head()

for year in df['Traffic Year'].unique():
    print 'In year %d, there are %d unique segments' %(year, len(df[df['Traffic Year'] == year]))

db = pd.read_csv('../FinalDatasets/sid.csv')

def getUID(row):
    rd = row['SR']
    bg = row['bs']
    end = row['es']
    res = db[((db.SR==rd) & ((db.bs == bg) & (db.es == end)))].UID
    if list(res) == []:
        return None
    else:
        return list(res)[0]

def roundme(n):
    return int(n*10)

df.columns = [                                u'SR',
                                      u'RRT',
                                      u'RRQ',
                                u'BA',
                                  u'EA',
                               u'Begin SRMP',
                                 u'End SRMP',
                   u'Direction To Inventory',
                             u'Traffic Year',
                                     u'AADT',
                            u'Single Trucks',
                            u'Double Trucks',
                             u'Train Trucks',
                     u'Peak Hour Percentage',
              u'Peak Hour Directional Split',
       u'Traffic Year ESALs _One Direction_',
                           u'Percent Trucks']

df.loc[:, 'bs'] = df['BA'].apply(roundme)
df.loc[:, 'es'] = df['EA'].apply(roundme)

su = df.drop_duplicates(subset = ['SR', 'BA', 'EA'])

su.loc[:, 'UID'] = su.apply(getUID, axis = 1)
print len(su[np.isnan(su.UID)])
print len(su)

su_wid = su[~np.isnan(su.UID)]
su2 = su[np.isnan(su.UID)].sort(columns = ['SR', 'BA', 'EA'])

su_ids = np.array(su_wid.UID)
print len(su_wid), len(su2)

def fillUID(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    begin = row['bs']
    if (end - bg > 0.05) & (end - bg < 0.15):
        res = db[((db.SR==rd) & (db.bs == begin))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in su_ids: #  change su_ids to a list and just append x to it if not found in ids.
                return x
            return None

su2.loc[:, 'UID'] = su2.apply(fillUID, axis = 1)

su_w2 = su2[~np.isnan(su2.UID)]
su3 = su2[np.isnan(su2.UID)]
print len(su3)

def fillUID2(row):
    rd = row['SR']
    bg = row['BA']
    end = row['EA']
    begin = row['bs']
    if (end - bg > 0.01):
        res = db[((db.SR==rd) & (db.bs == int((bg*10 + end*10)/2)))].UID
        if list(res) == []:
            return None
        else:
            x = list(res)[0]
            if x not in su_ids:
                return x
            return None

su3.loc[:, 'UID'] = su3.apply(fillUID2, axis = 1)

su_w3 = su3[~np.isnan(su3.UID)]
su4 = su3[np.isnan(su3.UID)]
su_ids = np.concatenate((su_ids,np.array(su_w3.UID)),axis=0)

su_final = pd.concat([su_wid, su_w2, su_w3])
su_ds = su_final.drop_duplicates(subset = ['UID'])
print len(su_ds)

su2 = su_ds[['UID', 'SR', 'EA', 'BA']]

df2    = pd.merge(su2, df, how = 'left', left_on = ['SR', 'BA', 'EA'], right_on = ['SR', 'BA', 'EA'])
# df_wid = df2.dropna(subset = ['UID'])
df_wid = df2.dropna(subset = ['Traffic Year'])
print len(df2), len(df_wid), len(df)

df_wid.head()

df_wid.columns = [                             u'UID_x',
                                       u'SR',
                                       u'EA',
                                       u'BA',
                                      u'RRT',
                                      u'RRQ',
                               u'Begin SRMP',
                                 u'End SRMP',
                   u'Direction To Inventory',
                             u'Traffic Year',
                                     u'AADT',
                            u'Single Trucks',
                            u'Double Trucks',
                             u'Train Trucks',
                     u'Peak Hour Percentage',
              u'Peak Hour Directional Split',
       u'Traffic Year ESALs _One Direction_',
                           u'Percent Trucks',
                                       u'bs',
                                       u'es',
                                    u'UID']

data = df[[u'SR',
                                       u'EA',
                                       u'BA',
                   u'Direction To Inventory',
                             u'Traffic Year',
                                     u'AADT',
                            u'Single Trucks',
                            u'Double Trucks',
                             u'Train Trucks',
                     u'Peak Hour Percentage',
              u'Peak Hour Directional Split',
       u'Traffic Year ESALs _One Direction_',
                           u'Percent Trucks',
                                    u'UID']]

data.head()

fdb = pd.read_csv('../modelDatasets/FWD.csv')

print len(fdb)

data.to_csv('../modelDatasets/roughnessModelDataset.csv')
