==> 16. Refining dataset for creating efficient roughness-cracking model.ipynb <==
import pandas as pd, numpy as np

d1 = pd.read_csv('../modelDatasets/roughnessModelDataset.csv')

d2 = pd.read_csv('../modelDatasets/fullDataset.csv')

print len(d2)
d2.head()

d1.head()

d1.columns

d1s = d1[['UID', 'Traffic Year', 'AADT', 'Traffic Year ESALs _One Direction_', 'Percent Trucks', 
          u'Single Trucks', u'Double Trucks', u'Train Trucks']]

d1s.columns = ['UID', 'tYear', 'AADT', 'ESAL', 'pTrucks', u'SingleTrucks', u'DoubleTrucks', u'TrainTrucks']

ds = pd.merge(d1s, d2, how = 'left', left_on = ['UID', 'tYear'], right_on = ['UID', 'sYear'])

print len(ds), len(d1s), len(d2)

df = ds.dropna(subset = ['ESAL', 'UID'])
df = df.drop_duplicates(subset = ['UID', 'tYear'])

df.to_csv('cp255_A8.csv')

print len(df), len(ds)

df.head()

data = df.dropna(subset = ['UID'])

print len(data.UID.unique())

ds = data.sort(columns = ['UID', 'tYear'])

ds.head()

# for year in ds.tYear.unique():
#     print '%d: %d unique segments' %(year, len(ds[ds.tYear == year]))
# ds[ds.UID ==1].tYear
# n = 0
# for x in ds.UID.unique():
#     temp = ds[ds.UID == x]
#     if len(temp) == 10:
#         n+=1
# print n

# Next step is to identify the rows that will be useful for perfoming the 1st Difference Regression 
After this stage, I will be able to model roughness with cracking and other covariates and possibly get more efficient parameter estimates. The work is to be continued shortly

# test = ds[(ds.UID == 146) & (ds.tYear == 2010) & (ds.tYear == 2009) ]
# print len(test)

#Thought I was gonna do diff in diffs but it didnt work as planned
# year_list = []
# for x in ds.UID.unique():
#     temp = ds[(ds.UID == x)]
#     years = list(np.array(temp.tYear))
#     toAppend = years[1:] + [0]
#     year_list += toAppend
# ds.loc[:, 'dYear'] = year_list
# ds.loc[:, 'year_diff'] = ds.apply(lambda x: x['dYear'] - x['tYear'], axis = 1)

# Now using whole dataset

# ds = d1.copy()

print len(ds[ds.Alli.isnull()==True])
print len(ds[(ds.Long.isnull()==0) & (ds.Alli.isnull()==1)])
print len(ds[(ds.Long.isnull()==1) & (ds.Alli.isnull()==0)])
print len(ds[(ds.Long.isnull()==0) & (ds.Alli.isnull()==0)])
print len(ds)
print len(ds[ds.PSC.isnull() ==1])

subdf = ds[((ds.Long.isnull()==0) | (ds.Alli.isnull()==0)) & (ds.PSC.isnull() ==0)]
print len(subdf)

# For model PSC = c + a * Alli + b * Long
c = 79.283
a = -0.121
b = -0.038
woAlli = subdf[subdf.Alli.isnull() == 1]
woLong = subdf[subdf.Long.isnull() == 1]

woAlli.Alli = woAlli.PSC.sub(woAlli.Long.mul(b)).sub(c).mul(1/a)
woLong.Long = woLong.PSC.sub(woLong.Alli.mul(a)).sub(c).mul(1/b)

woAlli.Alli = woAlli.apply(lambda x: 0 if x['Alli'] < 0 else x['Alli'], axis = 1)
woLong.Long = woLong.apply(lambda x: 0 if x['Long'] < 0 else x['Long'], axis = 1)

woNulls = subdf[(subdf.Alli.isnull() == 0) & (subdf.Long.isnull() == 0)]
print len(woNulls)

model_ds = pd.concat([woNulls, woAlli, woLong])
print len(model_ds) == len(subdf)

#Changing IRI units from inches/mile to meters/Km
model_ds.loc[:, 'IRIsi'] = model_ds.IRI * 0.08333333 * 0.189394

#Changing feet to miles for the respective segments
model_ds.loc[:, 'Alli'] = 0.000189394 * model_ds.Alli
model_ds.loc[:, 'Long'] = 0.000189394 * model_ds.Long

# Determining percent cracking as the total length of cracks divided by survey section
model_ds.loc[:, 'pAlli'] = model_ds.Alli.div(model_ds.loc[:, 'Lane Miles']).mul(100)
model_ds.loc[:, 'pLong'] = model_ds.Long.div(model_ds.loc[:, 'Lane Miles']).mul(100)

nds = model_ds[(model_ds.IRIsi < 8) & (model_ds.pAlli < 80) & (model_ds.pLong < 200)]
print len(nds)*100.0/len(model_ds)

nds['IRIsi'].head()

def getDiff(db, variable, newName):
    val_list = []
    for x in db.UID.unique():
        temp = db[(db.UID == x)]
        values = list(np.array(temp[variable]))
        toAppend = values[1:] + [0]
        val_list += toAppend
    db.loc[:, newName] = val_list

ndf = nds.sort(columns = ['UID', 'tYear'])

for item in ['Rutting', 'IRIsi', 'pAlli', 'pLong', 'cumPPT']:
    newName = 'd'+item
    getDiff(ndf, item, newName)

ndf.head()
