==> 17. Adding weather variables to new roughness model dataset.ipynb <==
import pandas as pd
import numpy as np

dat = pd.read_excel('../modelDatasets/CIMdataset2.xlsx')

ppt = pd.read_csv('../FinalDatasets/ppt.csv')
maxt = pd.read_csv('../FinalDatasets/MaxT.csv')
mint = pd.read_csv('../FinalDatasets/MinT.csv')
esals =pd.read_csv('../FinalDatasets/ESALs.csv')
rdat = pd.read_csv('../modelDatasets/rmdata1988_1999.csv')

#Rehabilitation years
Ryrs = rdat[['UID', 'Major Rehab Year', 'Prior Major Rehab Year']]
Ryrs.columns = ['UID', 'MRYear', 'PMRYear']
Ryrs = Ryrs.drop_duplicates(subset = 'UID')
Ryrs.index = Ryrs.UID

import math
def removeNan(row):
    val = row['PMRYear']
    if math.isnan(val):
        return 1900
    return val

Ryrs.loc[:, 'PriorRehab'] = Ryrs.apply(removeNan, axis = 1)

Ryrs[Ryrs.PriorRehab==1900].head()

print len(np.unique(np.array(esals.ID)))*33

selected_ids = np.unique(np.array(ppt.ID))
esals = esals[esals.ID.isin(selected_ids)]

esals = esals.drop_duplicates(subset = ['ID'])
ppt  = ppt.drop_duplicates(subset = ['ID'])
mint = mint.drop_duplicates(subset = ['ID'])
maxt = maxt.drop_duplicates(subset = ['ID'])

print len(esals)
print len(ppt)
print len(mint)
print len(maxt)
print len(Ryrs)

def addPrevRehabYear(row):
    ix = row['ID'] #index
    return Ryrs.loc[ix, 'PriorRehab']

def addRehabYear(row):
    ix = row['ID'] #index
    return Ryrs.loc[ix, 'MRYear']

def addRehabYears(df):
    df.loc[:, 'PriorMRyr'] = df.apply(addPrevRehabYear, axis =1)
    df.loc[:, 'MajorRyr'] = df.apply(addRehabYear, axis =1)

addRehabYears(esals)
addRehabYears(ppt)
addRehabYears(mint)
addRehabYears(maxt)

esals.head()

def cumVal(row):
    ix = row['ID']
    syear = row['Year']
    l = np.array(row[['1981', '1982', '1983', '1984', '1985', '1986', '1987', '1988', '1989', '1990', 
                      '1991', '1992', '1993', '1994', '1995', '1996', '1997', '1998', '1999', '2000',
                     '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008', '2009', '2010', 
                     '2011', '2012', '2013']])
    pmr = int(row['PriorMRyr'])
    if pmr > 1981:
        p1 = pmr - 1981
    else:
        p1 =0
    mr  = int(row['MajorRyr'])
    if (mr > 1981) & (mr < 2014):
        p2 = mr - 1981
    else:
        p2 =0
        
    if (p1 ==0) & (p2==0):
        res = list(np.cumsum(l))
        
    elif p1 == 0:
        l1 = list(np.cumsum(l[0:p2]))
        l2 = list(np.cumsum(l[p2:]))
        res = l1 + l2
    else:
        l1 = list(np.cumsum(l[0:p1]))
        l2 = list(np.cumsum(l[p1:p2]))
        l3 = list(np.cumsum(l[p2:]))
        res = l1 + l2 + l3
        
    return ix, syear, res

distributeResultToColumns = lambda x: pd.Series(([i for i in x[:2]] + [j for j in x[2:][0]]))
colNames = esals.columns[1:-2]
def getCumVarDataset(df):
    rdf = df.apply(cumVal, axis = 1)
    rdf = pd.DataFrame(rdf)
    rdf.columns = ['Results']
    newdf = rdf['Results'].apply(distributeResultToColumns)
    output = newdf[range(35)]
    output.columns = colNames
    return output

def cmean(row): #function determines the cumulative variables in a row
    ix = row['ID']
    syear = row['Year']
    l = np.array(row[['1981', '1982', '1983', '1984', '1985', '1986', '1987', '1988', '1989', '1990', 
                      '1991', '1992', '1993', '1994', '1995', '1996', '1997', '1998', '1999', '2000',
                     '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008', '2009', '2010', 
                     '2011', '2012', '2013']])
    pmr = int(row['PriorMRyr'])
    if pmr > 1981:
        p1 = pmr - 1981
    else:
        p1 =0
    mr  = int(row['MajorRyr'])
    if (mr > 1981) & (mr < 2014):
        p2 = mr - 1981
    else:
        p2 =0
        
    if (p1 ==0) & (p2==0):
        r = np.cumsum(l)
        d = np.arange(len(r))+1
        res = list(r/d)
        
    elif p1 == 0:
        l1 = np.cumsum(l[0:p2])
        d1 = np.arange(len(l1))+1
        l2 = np.cumsum(l[p2:])
        d2 = np.arange(len(l2))+1
        meanl1 = l1/d1
        meanl2 = l2/d2
        res = list(meanl1) + list(meanl2)
    else:
        l1 = np.cumsum(l[0:p1])
        d1 = np.arange(len(l1))+1
        l2 = np.cumsum(l[p1:p2])
        d2 = np.arange(len(l2))+1
        l3 = np.cumsum(l[p2:])
        d3 = np.arange(len(l3))+1
        meanl1 = l1/d1
        meanl2 = l2/d2
        meanl3 = l3/d3
        res = list(meanl1) + list(meanl2) + list(meanl3)
        
    return ix, syear, res

distributeResultToColumns = lambda x: pd.Series(([i for i in x[:2]] + [j for j in x[2:][0]]))
colNames = esals.columns[1:-2]
def cmeanVarDataset(df):
    rdf = df.apply(cmean, axis = 1)
    rdf = pd.DataFrame(rdf)
    rdf.columns = ['Results']
    newdf = rdf['Results'].apply(distributeResultToColumns)
    output = newdf[range(35)]
    output.columns = colNames
    return output

cumPpt   = getCumVarDataset(ppt)
cumEsals = getCumVarDataset(esals)
meanMaxT = cmeanVarDataset(maxt)
meanMinT = cmeanVarDataset(mint)

meanMinT.head()

cumPpt.to_csv('../FinalDatasets/cumPPT.csv')
cumEsals.to_csv('../FinalDatasets/cumEsals.csv')
meanMaxT.to_csv('../FinalDatasets/meanMaxT.csv')
meanMinT.to_csv('../FinalDatasets/meanMinT.csv')
