==> 18. Adding Speed Limits to MC dataset.ipynb <==
import pandas as pd
import numpy as np
import xlsxwriter

df1 = pd.read_excel('../DataFiles/Preservation Units.xlsx')

df1.head()

df2 = pd.read_csv('../tables/Preservation Units.txt', header = None)

len(df1) == len(df2)

df2.columns = df1.columns

df2.head()

df2[df2.SR==2]['Begin ARM'].describe()

df2[df2.SR==2]['Speed Limit'].unique()

print len(df2)

firstList = np.arange(df2[df2.SR==2]['Begin ARM'].max())
firstIndex = np.repeat(2, len(firstList))
ds = pd.DataFrame(firstList, firstIndex)
for sr in df2.SR.unique()[1:]:
    mx = df2[df2.SR==sr]['Begin ARM'].max()
    milepost = np.arange(mx)
    index = np.repeat(sr, len(milepost))
    sdf = pd.DataFrame(milepost, index)
    ds = pd.concat([ds, sdf])

ds.columns = ['milepost']
ds.tail(9)

df2.columns

df = df2[['SR', 'Begin ARM', 'Speed Limit']]

df.head(10)

ds['SR'] = ds.index

ds.index = np.arange(len(ds))

ds.head()

def getSpeed(row):
    sr = row['SR']
    mp = row['milepost']
    result = np.array(df[(df.SR == sr) & (df['Begin ARM'] >= mp) & (df['Begin ARM'] < (mp+1))]['Speed Limit'])
    if len(result) == 0:
        return None
    else:
        return result[0]

ds['SpeedLimit'] = ds.apply(getSpeed, axis = 1)

ds.head(10)

ds['ID'] = ds.index

def ImputeMissingSpeed(row):
    speed = row['SpeedLimit']
    mile  = row['milepost']
    index = int(row['ID'])
    if np.isnan(speed) & (mile == 0):
        ds.loc[index, 'SpeedLimit'] = 0
        return None
    elif np.isnan(speed):
        ds.loc[index, 'SpeedLimit'] = ds.loc[(index-1), 'SpeedLimit']
        return None
    else:
        return None

result = ds.apply(ImputeMissingSpeed, axis = 1)

ds.head(20)

print len(ds[ds['SpeedLimit'] == 0])

ds.SpeedLimit.unique()

ds2 = pd.read_csv('../modelDatasets/ndata.csv')

ds2.columns

#Importing the dataframe with unique ids
idf = pd.read_csv('../FInalDatasets/Contracts.csv')
idf = idf.drop_duplicates(subset = ['UID'])

print len(idf)
idf.head()

print len(ds2[ds2.UID.isin(idf.UID)])
print len(ds2)

sdf = idf[['SR', 'BA', 'UID']]
sdf.columns = ['SR', 'BMP', 'UID']

data = pd.merge(ds2, sdf, how = 'left', left_on = ['UID'], right_on = 'UID')

data.head()

def getSpeed(row):
    sr = row['SR']
    mp = int(row['BMP'])
    req = np.array(ds[(ds.SR == sr) & (ds.milepost==mp)]['SpeedLimit'])
    if len(req) == 0:
        return None
    return req[0]

data['SpeedLimit'] = data.apply(getSpeed, axis = 1)

data.head()

data.SpeedLimit.unique()

print len(data[data.SpeedLimit.isnull()==1])

dataset = data.sort(columns = ['SR', 'BMP'])

dataset.index = np.arange(len(data))
dataset['idx'] = np.arange(len(data))

dataset.head()

def ImputeSpeed(row):
    speed = row['SpeedLimit']
    idx   = row['idx']
    sr    = row['SR']
    mp    = row['BMP']
    if np.isnan(speed) & (mp == 0):
        dataset.loc[idx, 'SpeedLimit'] = 0
        return None
    elif np.isnan(speed):
        dataset.loc[idx, 'SpeedLimit'] = dataset.loc[(idx-1), 'SpeedLimit']
        return None     
    else:
        return None

res = dataset.apply(ImputeSpeed, axis = 1)

print len(dataset[dataset.SpeedLimit ==0])

dataset.SpeedLimit.unique()

dataset.columns

rdat = dataset[[ u'UID', u'IsBridge', u'Lane Surface Type', u'LaneMiles',
       u'Region_x', u'FFClass', u'SurfaceTypeCode', u'STDescription',
       u'TreatBaseCode', u'TreatBaseDescription', u'TBThick', u'untrthick',
       u'Y_ESAL', u'MajorRehabYear', u'Cum_ESAL', u'SURFTHK', u'ULT',
       u'TSfThick', u'tmin', u'tmax', u'cumESALs', u'IRIsi', u'Alli', u'Long',
       u'E_Alli', u'E_Long', u'prep', u'MinT', u'MaxT', u'FTCycles',
       u'actbthick', u'pctbthick', u'ftprep', u'BA', u'AA', u'SR', u'BMP',
       u'SpeedLimit']]

rdat.to_csv('../modelDatasets/MCdata.csv')

speeds = np.sort(dataset.SpeedLimit.unique())

speeds
