==> 19. Adding weather-variable means and std to the MC dataset.ipynb <==
import pandas as pd, numpy as np

d1 = pd.read_csv('../modelDatasets/weather_vars_stats2.csv')
d2 = pd.read_csv('../../Research Data/Rmodels/data/mdata.csv')
d3 = pd.read_csv('../../Research Data/Rmodels/data/mcdata.csv')

d2.columns

d3.columns

print len(d2) == len(d3)
d2.head()

d4 = d3[['UID', 'SpeedLimit']]

d5 = pd.merge(d2, d4, how = 'left', left_on = ['UID'], right_on = 'UID')
print len(d5) == len(d2)

d5.head()

d2 = d5.copy()

print len(d2[d2.UID.isin(d1.UID)])*100.0/len(d2)

d3 = pd.merge(d2,d1, how = 'left', left_on = 'UID', right_on = 'UID')
print len(d3) == len(d2)

d3.head()

d3.columns

d3[['UID', 'Lane Surface Type', 'SURFTHK', 'ULT']].head()

ds = d3[[u'UID', u'IsBridge', u'Lane Surface Type', u'LaneMiles', u'Region_x',
       u'FFClass', u'SurfaceTypeCode', u'STDescription', u'TreatBaseCode',
       u'TreatBaseDescription', u'TBThick', u'untrthick', u'Y_ESAL',
       u'MajorRehabYear', u'Cum_ESAL', u'SURFTHK', u'ULT', u'TSfThick',
       u'tmin', u'tmax', u'cumESALs', u'IRIsi', u'Alli', u'Long', u'E_Alli',
       u'E_Long', u'prep', u'MinT', u'MaxT', u'FTCycles', u'actbthick',
       u'pctbthick', u'ftprep', u'BA', u'AA', u'AADT', u'ptrucks', u'Trucks',
       u'PCars', u'VehMiles', u'TruckMiles', u'SpeedLimit',
       u'ESALmean', u'ESALstd', u'PPTmean', u'PPTstd', u'MINTmean', u'MINTstd',
       u'MAXTmean', u'MAXTstd', u'FTCmean', u'FTCstd']]

ds['SpeedLimit'] = ds.SpeedLimit.astype(int)

ds.SpeedLimit.head()

ds.to_csv('../../Research Data/Rmodels/data/mcsdataset.csv')

# working on fuel efficiency dataset

df = pd.read_csv('../../Research Data/Rmodels/data/VehFE2.csv')

df.head()

df['IRI'] = df.IRI.mul(100)

df.columns

feds = df[[u'IRI', u'PC25', u'PT25', u'PC30', u'PT30', u'PC35',
       u'PT35', u'PC40', u'PT40', u'PC45', u'PT45', u'PC50', u'PT50', u'PC55',
       u'PT55', u'PC60', u'PT60', u'PC65', u'PT65', u'PC70', u'PT70']]

feds.head()

feds['IRI'] = feds.IRI.astype(int)

feds.IRI.head()

feds.to_csv('../../Research Data/Rmodels/data/VehFuelE.csv') # Vehicle Fuel Efficiency dataset
