==> 20. Roughness-cracking model dataset changed temp from degC to degF.ipynb <==
import pandas as pd, numpy as np

ds = pd.read_csv("~/Desktop/PhD_Research/Rmodels/data/rmData1115.csv")

print len(ds)
ds.head()

print ds.mint.min()

ds['mint'] = ds.mint.mul(9.0/5).add(32)

print ds.mint.min()

ds['maxt'] = ds.maxt.mul(9.0/5).add(32)

ds.head()

ds.columns

dt = pd.read_csv('../modelDatasets/fullDataset.csv')

dt.columns

dt2 = dt[[ u'SR', u'Begin ARM', u'End ARM', 'UID', u'sYear']]

dt2 = dt2.drop_duplicates(subset = 'UID')

dt2 = dt2[(dt2.sYear >= 1988)& (dt2.sYear < 2000)]

print len(dt2)
print len(ds[ds.UID.isin(dt2.UID)])

rdt = pd.merge(ds, dt2, how='left', left_on = 'UID', right_on = 'UID')
print len(rdt)

rdt.head()

rdf = rdt.sort(columns = ['UID', 'sYear_x'])

rdf = rdf.drop_duplicates(subset = ['UID', 'sYear_x'])

rdf.head(10)

selected_IDs = []
for sr in rdf.SR.unique():
    temp = rdf[rdf.SR == sr]
    min_id = temp.UID.min()
    max_id = temp.UID.max()
    n = int((max_id - min_id)/10.0)
    for m in range(n):
        if (m%2 == 0):
            p = np.array(temp.UID[(temp.UID >= (min_id+m*10)) & (temp.UID < (min_id+ (m+1)*10))])
            if len(p) > 0:
                s = np.random.choice(p, 1)[0] # select item in that range
                selected_IDs.append(s)

print len(selected_IDs)
selected_IDs[0:20]

data = rdf[rdf.UID.isin(selected_IDs)]

print len(data)
print len(ds)

data.columns

data = data[[u'UID', u'sYear_x', u'Rutting', u'IRI', u'PSC', u'Alli',
       u'Long', u'Is Bridge', u'Lane Miles', u'Segment Lane Miles',
       u'ESALs Since Major Rehab', u'sfThick', u'Tthick', u'TSfThick',
       u'STcode', u'cumPPT', u'meanMinTemp', u'meanMaxTemp', u'cumESALs',
       u'cumFTC', u'pAlli', u'pLong', u'IRIsi', u'esals', u'ppt', u'mint',
       u'maxt', u'ICAO', u'ftc', u'SR', u'Begin ARM', u'End ARM']]

data.columns = [u'UID', u'sYear', u'Rutting', u'IRI', u'PSC', u'Alli',
       u'Long', u'Is Bridge', u'Lane Miles', u'Segment Lane Miles',
       u'ESALs Since Major Rehab', u'sfThick', u'Tthick', u'TSfThick',
       u'STcode', u'cumPPT', u'meanMinTemp', u'meanMaxTemp', u'cumESALs',
       u'cumFTC', u'pAlli', u'pLong', u'IRIsi', u'esals', u'ppt', u'mint',
       u'maxt', u'ICAO', u'ftc', u'SR', u'Begin ARM', u'End ARM']

data.to_csv('../modelDatasets/rcModelData1215.csv')

# ds.to_csv('../modelDatasets/rmData1215.csv')
