==> 21. Adding coordinates to final roughness-cracking model dataset.ipynb <==
import pandas as pd, numpy as np

ds = pd.read_csv('../modelDatasets/rcModelData1215.csv')

print len(ds)
ds.head()

cd = pd.read_csv('../FinalDatasets/RCsegCoord.csv')

cd.head()

df = ds.drop_duplicates(subset = ['UID'])

print len(df[df.UID.isin(cd.UID)]) == len(df)

mds = pd.merge(df,cd, how = 'left', left_on = 'UID', right_on = 'UID')
print len(mds) == len(df)

rds = mds[['UID', 'Latitude', 'Longitude']]
rds.head()

print len(rds)
rds.columns = ['UID', 'latitude', 'longitude']

rds.to_csv('../modelDatasets/rm_seg_coords.csv') # Roughness model
