==> 23. Converting °C to °F for temperature variables.ipynb <==
import pandas as pd, numpy as np

ds = pd.read_csv("../../Research Data/Rmodels/data/mcsdataset.csv")

ds.head()

ds.columns

ds[['tmax', 'tmin', 'MinT', 'MaxT']].describe()

ds['tmin'] = ds.tmin.mul(9.0/5).add(32)
ds['tmax'] = ds.tmax.mul(9.0/5).add(32)

ds['MINTmean'] = ds.MINTmean.mul(9.0/5).add(32)
ds['MINTstd'] = ds.MINTstd.mul(9.0/5).add(32)
ds['MAXTmean'] = ds.MAXTmean.mul(9.0/5).add(32)
ds['MAXTstd'] = ds.MAXTstd.mul(9.0/5).add(32)

ds.to_csv("../../Research Data/Rmodels/data/mcsdataset2.csv")
